Remove useless casts in similar/arch/ogl/ogl.cpp

+++ similar/arch/ogl/ogl.cpp
@@ -533,8 +533,8 @@ void g3_draw_line(const g3s_point &p0,const g3s_point &p1, const uint8_t c)
 	GLfloat color_r, color_g, color_b;
 	GLfloat color_array[] = { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0 };
 	GLfloat vertex_array[] = {
-		static_cast<GLfloat>(f2glf(p0.p3_vec.x)), static_cast<GLfloat>(f2glf(p0.p3_vec.y)), static_cast<GLfloat>(-f2glf(p0.p3_vec.z)),
-		static_cast<GLfloat>(f2glf(p1.p3_vec.x)), static_cast<GLfloat>(f2glf(p1.p3_vec.y)), static_cast<GLfloat>(-f2glf(p1.p3_vec.z))
+		f2glf(p0.p3_vec.x), f2glf(p0.p3_vec.y), -f2glf(p0.p3_vec.z),
+		f2glf(p1.p3_vec.x), f2glf(p1.p3_vec.y), -f2glf(p1.p3_vec.z)
 	};
   
 	ogl_client_states<int, GL_VERTEX_ARRAY, GL_COLOR_ARRAY> cs;
